implement flash and zoom in functionality.

+++ app/src/main/java/com/example/ticketpartner/feature_scan_module/feature_qr_scan/data/repository/QrScanRepositoryImpl.kt
@@ -33,7 +33,7 @@ class QrScanRepositoryImpl @Inject constructor(private val qrScanDataSource: QrS
     }
 
     override suspend fun getScanCheckedInResponse(
-        checkedOrderIdList: ArrayList<Int>,
+        checkedOrderIdList: ArrayList<Long>,
         orderId: String
     ): ScanCheckedInResponse {
         return qrScanDataSource.getScanCheckedInResponse(checkedOrderIdList, orderId)
